split out FFICXX.Generate.Name from FFICXX.Generate.Code.Primitive
